Make it compatible with browser without css animation an jquery.easing
If a user didn't include `jquery.easing.1.3.js` in their script and use a browser without CSS animation like IE8, a user would receive `Object doesn't support property or method 'easeInOutExpo'` from jQuery and the slider wouldn't run. 

It happens because the liquid slider use the default animation `easeInOutExpo` and since the user didn't include `jquery.easing.1.3.js` the jquery returns an error when trying to use the easing. 

To fix this we can check if jQuery support the easing, if it doesn't we fallback to `swing` easing which is supported by default.

This is not a bug fix, it is more an enhancement to remove the strict requirement of `jquery.easing.1.3.js` script.